smart_farming_recommender/requirements.txt
Django
djangorestframework
python-dotenv
langchain-google-genai
langchain-community
langchain-core
langgraph
Pillow
numpy
sentence-transformers
tavily-python